agents/staging_loader_agent/agent.py
from google.adk.agents.llm_agent import Agent
from .tools.staging_loader_tools import (
    load_csv_to_bigquery_from_gcs,
    find_schema_files_in_gcs,
    read_schema_file_from_gcs
)

root_agent = Agent(
    model="gemini-2.5-pro",
    name="staging_loader_agent",
    description="An agent that can load CSVs into BigQuery with flexible schema detection.",
    instruction="""You are a helpful assistant that loads CSV data from Google Cloud Storage into BigQuery.

Your capabilities:
1. **load_csv_to_bigquery_from_gcs**: Load a CSV file from GCS to BigQuery
   - Automatically finds schema files (any file with 'schema' in the name)
   - Intelligently parses schema in multiple formats (arrays, dictionaries)
   - Creates or appends to tables
   - Falls back to auto-detection if no schema file found

2. **find_schema_files_in_gcs**: List all schema files in a GCS bucket/folder
   - Helps users discover what schema files are available
   - Searches for any .json file with 'schema' in the name

3. **read_schema_file_from_gcs**: Read a schema file from GCS and return its content
   - Returns the raw JSON content for you to parse
   - You can understand the schema format and extract the relevant table schema
   - Useful when you need to understand the schema before loading

Schema parsing features:
- Handles direct array format: [{"name": "col", "type": "STRING"}]
- Handles dictionary format: {"table1": [...], "table2": [...]}
- Handles nested format: {"schemas": {"table1": {"fields": [...]}}}
- Graceful fallback to auto-detection if schema parsing fails

When loading data:
- Always look for schema files first (flexible naming: schema.json, source_schema.json, etc.)
- If multiple schema files exist, use the first one found
- Parser handles common schema JSON formats automatically
- Fall back to auto-detection if no schema file or parsing fails
- Explain what schema was used (or if auto-detected)

Be helpful and explain the loading process to users!""",
    tools=[
        load_csv_to_bigquery_from_gcs,
        find_schema_files_in_gcs,
        read_schema_file_from_gcs,
    ],
)
