agents/staging_loader_agent/requirements.txt
google-adk
google-cloud-bigquery
google-cloud-storage
pandas
python-dotenv
